Adds test application used in QR development
Test application used to debug the code being written for the pam module
ie the code that converts the raw QR code data into a String to be
displayed by the application.

Contains duplicate of the code in pam_qr.c for generating QR code string
at time of commit (subject to change, check commits). This code won't be
actively updated unless it is needed for testing and will probably
eventually be removed.