Refactored QR Code generation code
Moved QR code string generation to separate file which can be run as
it's own standalone program or used as a library to generate strings
from qr code data encoded with libqrencode. Could be spun into it's own
project but will live in this repo for now.

Added makefile rule to compile this new program and deleted old test
program in favour of printqr.