Adding parsing and usage of auth string from API
Now using string obtained from API to check whether the user is
authenticated or not. Currently just checking for "authenticated": true.
This is a proof of concept and will need worked on to make it secure.

Updated makfile and install script due to implementation and tidy ups in
other files in this commit. Also added some missing header files to some
c files.