Reimplemented converting QR code data to string
Reimplemented using ASCII blocks (which I thought I was already using
but apparently I was just colouring the terminal), which now allows it
to be printed when logging in via SSH fixing the major bug with this
software!!

Added rule to compile qr_to_string test. Not a proper test, yet (maybe),
just used to visually check the output.